+++ src/main/resources/archetype-resources/src/test/java/config/WebSecurityConfigurationAware.java
@@ -1,6 +1,8 @@
 package ${package}.config;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.context.event.ContextStartedEvent;
 import org.springframework.security.web.FilterChainProxy;
 
 import javax.inject.Inject;
@@ -12,10 +14,14 @@ public abstract class WebSecurityConfigurationAware extends WebAppConfigurationA
     @Inject
     private FilterChainProxy springSecurityFilterChain;
 
+    @Inject
+    private ApplicationEventPublisher publisher;
+    
     @BeforeEach
     public void before() {
         this.mockMvc = webAppContextSetup(this.wac).apply(springSecurity())
                 //.addFilters(this.springSecurityFilterChain)
                 .build();
+        publisher.publishEvent(new ContextStartedEvent(this.wac));
     }
 }